requirement.txt
paddleocr==2.6.1.3
paddlepaddle==2.6.2
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==10.3.0
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.40.2
protobuf==3.20.2